requirements.txt
customtkinter
pillow
qrcode
matplotlib
tkcalendar
schedule
